# docs/openapi/horizon-ric-rapp.yaml
openapi: 3.1.0
info:
  title: PreceptualAI Operator API
  description: "PreceptualAI rApp operator API. O-RAN.WG2 R1/A1/O1 conformant, TM Forum\
    \ ODA Production component. JWT bearer auth (HS256/RS256) per O-RAN.WG11 \xA7\
    6."
  version: 0.2.0
paths:
  /v1/state:
    get:
      summary: Get State
      operationId: get_state_v1_state_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAppStateInfo'
      security:
      - HTTPBearer: []
...
  /api/v1/state:
    get:
      summary: Get State
      operationId: get_state_api_v1_state_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAppStateInfo'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    RAppStateInfo:
      properties:
        state:
          type: string
          title: State
        rapp_id:
          type: string
          title: Rapp Id
        version:
          type: string
          title: Version
        uptime_seconds:
          type: number
          title: Uptime Seconds
        last_decision_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Decision At
      type: object
      required:
      - state
      - rapp_id
      - version
      - uptime_seconds
      title: RAppStateInfo
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Operator JWT (HS256 dev, RS256 prod). Audience: 'horizon-ric'.\
        \ Conforms to O-RAN.WG11 \xA76 token requirements."
security:
- BearerJWT: []
servers:
- url: http://localhost:8083
  description: Local
- url: https://horizon-ric.example.com
  description: Production